School Assessment, Food menu
I am new to coding and one of my first school assessment is to design and simulate a cashless tuckshop. Anyway I am trying to design a function which will:
1. Display the menu
2. Ask for what you would like
3. Check to see if the item is in the menu
4. if it is in the menu, then add it to your order, 
5. then print out the order and tell you how much it will cost
however I am getting an error and I do not know how to fix it
line 22, in FoodMenu
    order[menu].append(item)
TypeError: unhashable type: 'dict'


if anyone can improve the code you are more than welcome to.